feat: implement board view layout and routing
